ES-303 remove reliability issue

+++ engagementsdk/src/main/java/com/livelike/engagementsdk/widget/util/SwipeDismissTouchListener.java
@@ -117,7 +117,7 @@ public class SwipeDismissTouchListener implements View.OnTouchListener {
                 float absVelocityY = Math.abs(mVelocityTracker.getYVelocity());
                 boolean dismiss = false;
                 boolean dismissRight = false;
-                if (Math.abs(deltaX) > ((float) (mViewWidth / 2)) && mSwiping) {
+                if (Math.abs(deltaX) > mViewWidth / 2 && mSwiping) {
                     dismiss = true;
                     dismissRight = deltaX > 0;
                 } else if (mMinFlingVelocity <= absVelocityX && absVelocityX <= mMaxFlingVelocity
